Migrated UIViewcontroller extension.
